nr-fh-control: Consider FH capacity for full-duplex FH link

+++ model/nr-fh-control.h
@@ -135,7 +135,7 @@ class NR_EXPORT NrFhControl : public Object
      *        one of its UEs has data.
      * @param capacity The fronthaul capacity (in Mbps)
      */
-    void SetCellFhCapacity(uint32_t capacity);
+    void SetCellFhCapacity(uint64_t capacity);
 
     /**
      * @brief Set the overhead for dynamic modulation compression
@@ -339,7 +339,7 @@ class NR_EXPORT NrFhControl : public Object
     NrFhSchedSapProvider* m_fhSchedSapProvider; //!< FH Control -  SCHED SAP Provider
 
     enum FhControlMethod m_fhControlMethod;
-    uint32_t m_fhCapacity{
+    uint64_t m_fhCapacity{
         1000}; //!< the available FH capacity (in Mbps) for DL and UL (full-duplex FH link)
     uint8_t m_overheadDyn{32};    //!< the overhead (OH) for dynamic adaptation (in bits)
     uint8_t m_mcsTable{2};        //!< the MCS table
